refactor: use snprintf instead of sprintf

+++ src/application.hpp
@@ -1,12 +1,10 @@
 #pragma once
 
-#include <cstddef>
-
 class Application {
 public:
     ~Application();
     void start();
 
-    constexpr static size_t WIDTH = 800;
-    constexpr static size_t HEIGHT = 600;
+    constexpr static int WIDTH = 800;
+    constexpr static int HEIGHT = 600;
 };
